pyproject.toml
[project]
name = "classroom-optimizer"
version = "0.1.0"
description = "AI-powered classroom seating optimization tool for teachers"
authors = [
    {name = "Daniel", email = "your-email@example.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-adk",
    "ortools>=9.8.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.5.0",
    "python-dotenv",
    "python-multipart",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "isort",
    "flake8",
    "mypy",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.black]
line-length = 88
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[dependency-groups]
dev = [
    "black",
    "flake8",
    "isort",
    "mypy",
    "pytest",
]

